requirements.txt
discord.py
psycopg2
requests
aiohttp[speedups]

aiocron
python-dateutil
dateparser
pytz

emoji
fuzzywuzzy[speedup]
python-dotenv